Add Tailwind styling and new courses page
